requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.16.13
botocore==1.19.13
certifi==2020.11.8
Django==3.1.2
django-cors-headers==3.5.0
django-debug-toolbar==3.2
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-extensions==3.1.1
django-nine==0.2.3
django-redis==4.12.1
django-storages==1.10.1
django-webpack-loader==0.7.0
djangorestframework==3.12.1
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
gunicorn==20.0.4
isort==5.6.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
wrapt==1.12.1
